feat(nextjs): support building and serving production server
