+++ packages/next/src/schematics/application/files/.babelrc__tmpl__
@@ -1,0 +1,4 @@
+{
+  "presets": ["next/babel"],
+  "plugins": [<% if (style === 'styled-components') { %>["styled-components", { "pure": true, "ssr": true }]<% } %> ]
+}
