--- packages/next/src/schematics/application/files/environments/environment.prod.ts__tmpl__
@@ -1,3 +1,0 @@
-export const environment = {
-  production: true
-};
