+++ packages/next/src/schematics/application/files/pages/_document.tsx__tmpl__
@@ -1,0 +1,32 @@
+import React, { ReactElement } from 'react';
+import Document, { Head, Main, NextScript } from 'next/document';
+import { ServerStyleSheet } from 'styled-components';
+
+export default class CustomDocument extends Document<{ styleTags: ReactElement[] }> {
+  static getInitialProps({ renderPage }) {
+    const sheet = new ServerStyleSheet();
+
+    const page = renderPage(App => props =>
+      sheet.collectStyles(<App {...props} />)
+    );
+
+    const styleTags = sheet.getStyleElement();
+
+    return { ...page, styleTags };
+  }
+
+  render() {
+    return (
+      <html>
+        <Head>
+          <title>Welcome to <%= name %>!</title>
+          {this.props.styleTags}
+        </Head>
+        <body>
+          <Main />
+          <NextScript />
+        </body>
+      </html>
+    );
+  }
+}
