+++ packages/next/src/schematics/init/init.ts
@@ -1,18 +1,18 @@
 import { chain, noop } from '@angular-devkit/schematics';
 import { addDepsToPackageJson, addPackageWithInit } from '@nrwl/workspace';
-import {
-  nextVersion,
-  zeitNextCss,
-  zeitNextLess,
-  zeitNextSass,
-  zeitNextStylus
-} from '../../utils/versions';
+import { nextVersion } from '../../utils/versions';
 import { Schema } from './schema';
 import { setDefaultCollection } from '@nrwl/workspace/src/utils/rules/workspace';
+import {
+  reactDomVersion,
+  reactVersion
+} from '../../../../react/src/utils/versions';
 
 const updateDependencies = addDepsToPackageJson(
   {
-    next: nextVersion
+    next: nextVersion,
+    react: reactVersion,
+    'react-dom': reactDomVersion
   },
   {}
 );
