+++ packages/next/src/utils/versions.ts
@@ -6,3 +6,5 @@ export const zeitNextSass = '1.0.1';
 export const nodeSass = '4.13.1';
 export const zeitNextLess = '1.0.1';
 export const zeitNextStylus = '1.0.1';
+export const emotionServerVersion = '10.0.27';
+export const babelPluginStyledComponentsVersion = '1.10.7';
